pyproject.toml
[project]
name = "theatre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.3",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "python-dotenv>=1.0.1",
    "stripe>=11.3.0",
    "django-cors-headers>=4.6.0",
    "django-filter>=24.3",
    "pillow>=11.0.0",
    "django-multiselectfield>=0.1.13",
]
